backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-filter>=23.0
django-cors-headers>=4.0
PyMySQL>=1.1 
Pillow>=10.0
python-decouple>=3.8